+++ lib/members/getSpacesPropertyValues.ts
@@ -1,13 +1,13 @@
 import { prisma } from 'db';
+import { getAccessibleMemberPropertiesBySpace } from 'lib/members/getAccessibleMemberPropertiesBySpace';
 import { getCommonSpaceIds } from 'lib/members/getCommonSpaceIds';
 import { getSpaceMemberRoles } from 'lib/members/getSpaceMemberRoles';
-import { getVisibleMemberPropertiesBySpace } from 'lib/members/getVisibleMemberPropertiesBySpace';
 import type { CommonSpacesInput, MemberPropertyValuesBySpace } from 'lib/members/interfaces';
 import { getPropertiesWithValues, groupPropertyValuesBySpace } from 'lib/members/utils';
 
 export async function getSpacesPropertyValues ({ memberId, requestingUserId, spaceId }: CommonSpacesInput): Promise<MemberPropertyValuesBySpace[]> {
   const spaceIds = requestingUserId ? await getCommonSpaceIds({ spaceId, memberId, requestingUserId }) : [];
-  const visibleMemberProperties = await getVisibleMemberPropertiesBySpace({ spaceId: spaceIds, userId: requestingUserId });
+  const visibleMemberProperties = await getAccessibleMemberPropertiesBySpace({ spaceId: spaceIds, userId: requestingUserId });
   const memberPropertyIds = visibleMemberProperties.map(mp => mp.id);
 
   const memberPropertyValues = await prisma.memberPropertyValue.findMany({
